requirement.txt
python-telegram-bot==20.7
phonenumbers
gradio
nest_asyncio